refactor: remove temporary merge file
